Aggiornamento con modifiche alle funzioni per gestire diversi tipi di annotazione tramite LLM
